requirements.txt
Flask==2.0.1
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==8.4.0"
rembg==1.0.38"
